Processed review comments
