+++ README.md
@@ -363,18 +363,12 @@ from orchestrator.core.settings import AppSettings
 import db  # noqa: F401  Side-effects: registers CustomerTable with ALL_DB_MODELS
 import products  # noqa: F401  Side-effects
 import workflows  # noqa: F401  Side-effects
-from graphql_federation import CUSTOM_GRAPHQL_MODELS, custom_subscription_interface
 
 app = OrchestratorCore(base_settings=AppSettings())
-app.register_graphql(
-    subscription_interface=custom_subscription_interface,
-    graphql_models=CUSTOM_GRAPHQL_MODELS,
-)
 ```
 
 All other orchestrator code is referenced by importing the `products`,
-`workflows`, and `db` modules. The `graphql_federation` module provides
-customised GraphQL types and models. The application is started with:
+`workflows`, and `db` modules. The application is started with:
 
 ```shell
 uvicorn --host localhost --port 8080 wsgi:app
@@ -1515,7 +1509,7 @@ The following is required to facilitate GraphQL federation on top of WFO and oth
 - The other backend must also enable federation
   - NetBox: [`docker/netbox/Dockerfile`](docker/netbox/Dockerfile)
 - In both backends set a federation key on the GraphQL types to join
-  - WFO: [`graphql_federation.py`](graphql_federation.py)
+  - WFO: [`graphql_utils/federation.py`](graphql_utils/federation.py)
   - NetBox: [`docker/netbox/patch_federation.py`](docker/netbox/patch_federation.py)
 - Define the supergraph config with both backends
   - [`docker/federation/supergraph-config.yaml`](docker/federation/supergraph-config.yaml)
@@ -1651,7 +1645,7 @@ This example orchestrator demonstrates the local table approach. A `CustomerTabl
     ALL_DB_MODELS.extend([CustomerTable])
     ```
 
-3. **Override the `customer` field on `SubscriptionInterface` and `ProcessType`** using `override_class` (see `graphql_federation.py`). This is the recommended approach from orchestrator-core — it ensures all autogenerated subscription types inherit the override:
+3. **Override the `customer` field on `SubscriptionInterface` and `ProcessType`** using `override_class` (see `graphql_utils/resolvers.py`). This is the recommended approach from orchestrator-core — it ensures all autogenerated subscription types inherit the override:
 
     ```python
     from orchestrator.core.graphql.schemas.process import ProcessType
