+++ wsgi.py
@@ -18,7 +18,7 @@ from orchestrator.core.settings import AppSettings
 import db  # noqa: F401  Side-effects: registers CustomerTable with ALL_DB_MODELS
 import products  # noqa: F401  Side-effects
 import workflows  # noqa: F401  Side-effects
-from graphql_federation import CUSTOM_GRAPHQL_MODELS, custom_subscription_interface
+from graphql_utils import CUSTOM_GRAPHQL_MODELS, custom_subscription_interface
 
 app = OrchestratorCore(base_settings=AppSettings())
 app.register_graphql(
